clean price page file and made it another component and clean code
